+++ js/src/admin/components/Settings.js
@@ -1,9 +1,26 @@
 import app from "flarum/admin/app";
 import ExtensionPage from "flarum/admin/components/ExtensionPage";
 import UploadImageButton from 'flarum/admin/components/UploadImageButton';
+import isExtensionInstalled from "../helpers/isExtensionInstalled";
+import isExtensionActive from "../helpers/isExtensionActive";
+import icon from "flarum/common/helpers/icon";
 
 export default class Settings extends ExtensionPage {
 	content() {
+		/* Supported third party extensions' info */
+		const viewsExtension = {
+			id: 'flarumite-simple-discussion-views',
+			name: 'Flarumite Simple Discussion Views',
+			url: 'https://flarum.org/extension/flarumite/simple-discussion-views',
+		};
+		const blogExtension = {
+			id: 'v17development-blog',
+			name: 'Flarum Blog',
+			url: 'https://flarum.org/extension/v17development/flarum-blog',
+		};
+
+		const warningIcon = icon("fas fa-exclamation-triangle");
+
 		return (
 			<div className="DiscussionCardsSettings">
 				<div className="container">
@@ -119,12 +136,55 @@ export default class Settings extends ExtensionPage {
 							{app.translator.trans("walsgit_discussion_cards.admin.settings.general.otherOptions_info")}
 						</p>
 						<div className="Section">
-							<h4>If <a href="https://flarum.org/extension/flarumite/simple-discussion-views" target="_blank" rel="noopener noreferrer">Flarumite Simple Discussion Views</a> is installed & activated</h4>
+							<h4>
+								{app.translator.trans("walsgit_discussion_cards.admin.settings.general.showViews_title_start")}
+								<a href={viewsExtension.url} target="_blank" rel="noopener noreferrer">{viewsExtension.name}</a>
+								{app.translator.trans("walsgit_discussion_cards.admin.settings.general.showViews_title_end")}
+							</h4>
+							<div className="helpText">
+								<em>
+									{!isExtensionInstalled(viewsExtension.id) 
+									? + app.translator.trans("walsgit_discussion_cards.admin.settings.general.viewsExtension_notInstalled", {icon: warningIcon}) : 
+									!isExtensionActive(viewsExtension.id) 
+									? app.translator.trans("walsgit_discussion_cards.admin.settings.general.viewsExtension_notActivated", {icon: warningIcon}) :
+									''}
+								</em>
+							</div>
 							{this.buildSettingComponent({
 								type: "switch",
 								setting: "walsgit_discussion_cards_showViews",
 								label: app.translator.trans("walsgit_discussion_cards.admin.settings.general.showViews_label"),
 								help: app.translator.trans("walsgit_discussion_cards.admin.settings.general.showViews_help"),
+								disabled: isExtensionActive(viewsExtension.id) ? false : true,
+							})}
+							<hr></hr>
+							<h4>
+								{app.translator.trans("walsgit_discussion_cards.admin.settings.general.blogExtension_title_start")}
+								<a href={blogExtension.url} target="_blank" rel="noopener noreferrer">{blogExtension.name}</a>
+								{app.translator.trans("walsgit_discussion_cards.admin.settings.general.blogExtension_title_end")}
+							</h4>
+							<div className="helpText">
+								<em>
+									{!isExtensionInstalled(blogExtension.id) 
+									? + app.translator.trans("walsgit_discussion_cards.admin.settings.general.blogExtension_notInstalled", {icon: warningIcon}) : 
+									!isExtensionActive(blogExtension.id) 
+									? app.translator.trans("walsgit_discussion_cards.admin.settings.general.blogExtension_notActivated", {icon: warningIcon}) :
+									''}
+								</em>
+							</div>
+							{this.buildSettingComponent({
+								type: "switch",
+								setting: "walsgit_discussion_cards_useBlogImages",
+								label: app.translator.trans("walsgit_discussion_cards.admin.settings.general.useBlogImages_label"),
+								help: app.translator.trans("walsgit_discussion_cards.admin.settings.general.useBlogImages_help"),
+								disabled: isExtensionActive(blogExtension.id) ? false : true,
+							})}
+							{this.buildSettingComponent({
+								type: "switch",
+								setting: "walsgit_discussion_cards_useBlogSummary",
+								label: app.translator.trans("walsgit_discussion_cards.admin.settings.general.useBlogSummary_label"),
+								help: app.translator.trans("walsgit_discussion_cards.admin.settings.general.useBlogSummary_help"),
+								disabled: isExtensionActive(blogExtension.id) ? false : true,
 							})}
 						</div>
 						{this.submitButton()}